+++ src/SolverInterfaces/SolverInterfaces.jl
@@ -1,5 +1,7 @@
 module SolverInterfaces
 
+using LinearAlgebra
+
 using Gridap
 using Gridap.Helpers
 using Gridap.Algebra
@@ -15,11 +17,14 @@ include("GridapExtras.jl")
 include("SolverTolerances.jl")
 include("ConvergenceLogs.jl")
 include("SolverInfos.jl")
+include("NullSpaces.jl")
 
 export SolverVerboseLevel, SolverConvergenceFlag
 export SolverTolerances, get_solver_tolerances, set_solver_tolerances!
 export ConvergenceLog, init!, update!, finalize!, reset!, print_message, set_depth!
 
 export SolverInfo
+
+export NullSpace
 
 end