+++ test/LinearSolvers/seq/runtests.jl
@@ -5,3 +5,4 @@ include("IterativeSolversWrappersTests.jl")
 include("SmoothersTests.jl")
 include("GMGTests.jl")
 include("RichardsonLinearTests.jl")
+